+++ apps/ng-bootstrap-demo/src/app/app.component.ts
@@ -1,7 +1,7 @@
 import { Component, inject, ChangeDetectionStrategy} from '@angular/core';
 import { SlideUpDownAnimation } from '@mintplayer/ng-animations';
 import { Color } from '@mintplayer/ng-bootstrap';
-import { BsNavbarComponent, BsNavbarNavComponent, BsNavbarDropdownComponent, BsNavbarItemComponent, BsNavbarContentDirective, BsNavbarBrandComponent, BsExpandButtonDirective } from '@mintplayer/ng-bootstrap/navbar';
+import { BsNavbarComponent, BsNavbarNavComponent, BsNavbarDropdownComponent, BsNavbarItemComponent, BsNavbarContentDirective, BsNavbarBrandComponent, BsExpandButtonDirective, BsNavbarTriggerDirective } from '@mintplayer/ng-bootstrap/navbar';
 import { BsDropdownDividerDirective } from '@mintplayer/ng-bootstrap/dropdown-divider';
 import { BsNavbarTogglerComponent } from '@mintplayer/ng-bootstrap/navbar-toggler';
 import { ViewportScroller } from '@angular/common';
@@ -13,7 +13,7 @@ import { BOOTSTRAP_VERSION } from './providers/bootstrap-version.provider';
   selector: 'demo-bootstrap-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss'],
-  imports: [RouterOutlet, RouterLink, BsNavbarComponent, BsNavbarNavComponent, BsNavbarDropdownComponent, BsNavbarItemComponent, BsNavbarContentDirective, BsNavbarBrandComponent, BsExpandButtonDirective, BsDropdownDividerDirective, BsNavbarTogglerComponent, NavbarRouterLinkActiveDirective],
+  imports: [RouterOutlet, RouterLink, BsNavbarComponent, BsNavbarNavComponent, BsNavbarDropdownComponent, BsNavbarItemComponent, BsNavbarContentDirective, BsNavbarBrandComponent, BsExpandButtonDirective, BsNavbarTriggerDirective, BsDropdownDividerDirective, BsNavbarTogglerComponent, NavbarRouterLinkActiveDirective],
   animations: [SlideUpDownAnimation],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })