+++ apps/ng-bootstrap-demo/src/app/app.config.ts
@@ -1,9 +1,10 @@
 import { provideHttpClient, withFetch } from "@angular/common/http";
 import { ApplicationConfig, provideBrowserGlobalErrorListeners, provideZonelessChangeDetection } from "@angular/core";
 import { provideAnimations } from "@angular/platform-browser/animations";
-import { PreloadAllModules, provideRouter, withEnabledBlockingInitialNavigation, withPreloading, withInMemoryScrolling } from "@angular/router";
+import { PreloadAllModules, withEnabledBlockingInitialNavigation, withPreloading, withInMemoryScrolling } from "@angular/router";
 import { provideHighlightOptions } from 'ngx-highlightjs';
 import ngBootstrapJson from '@mintplayer/ng-bootstrap/package.json';
+import { provideNavigationLock, provideNavigationLockRouter } from '@mintplayer/ng-bootstrap/navigation-lock';
 import { GIT_REPO } from "./providers/git-repo.provider";
 import { BOOTSTRAP_VERSION } from "./providers/bootstrap-version.provider";
 
@@ -13,7 +14,7 @@ export const config: ApplicationConfig = {
         provideHttpClient(withFetch()),
         provideZonelessChangeDetection(),
         provideBrowserGlobalErrorListeners(),
-        provideRouter(
+        provideNavigationLockRouter(
             [
                 { path: '', loadChildren: () => import('./pages/pages.routes').then(m => m.ROUTES) },
             ],
@@ -22,8 +23,9 @@ export const config: ApplicationConfig = {
             withInMemoryScrolling({
                 scrollPositionRestoration: 'enabled',
                 anchorScrolling: 'enabled',
-            })
+            }),
         ),
+        provideNavigationLock(),
         provideHighlightOptions({
             fullLibraryLoader: () => import('highlight.js'),
             themePath: 'solarized-dark.css'