+++ apps/ng-bootstrap-demo/src/app/components/confirm-dialog/confirm-dialog.component.spec.ts
@@ -1,0 +1,40 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { MockComponent, MockDirective } from 'ng-mocks';
+import {
+  BsModalHostComponent, BsModalDirective, BsModalHeaderDirective,
+  BsModalBodyDirective, BsModalFooterDirective, BsModalCloseDirective,
+} from '@mintplayer/ng-bootstrap/modal';
+
+import { ConfirmDialogComponent } from './confirm-dialog.component';
+
+describe('ConfirmDialogComponent', () => {
+  let component: ConfirmDialogComponent;
+  let fixture: ComponentFixture<ConfirmDialogComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [
+        MockComponent(BsModalHostComponent),
+        MockDirective(BsModalDirective),
+        MockDirective(BsModalHeaderDirective),
+        MockDirective(BsModalBodyDirective),
+        MockDirective(BsModalFooterDirective),
+        MockDirective(BsModalCloseDirective),
+
+        // Unit to test (standalone)
+        ConfirmDialogComponent,
+      ],
+    })
+    .compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(ConfirmDialogComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});