+++ libs/mintplayer-ng-bootstrap/navbar/src/index.ts
@@ -4,6 +4,7 @@ export * from './navbar-nav/navbar-nav.component';
 export * from './navbar-dropdown/navbar-dropdown.component';
 export * from './navbar-item/navbar-item.component';
 export * from './navbar-content/navbar-content.directive';
+export * from './navbar-trigger/navbar-trigger.directive';
 export * from './expand-button/expand-button.directive';
 
 // Not for end-use