+++ libs/mintplayer-ng-bootstrap/navigation-lock/src/index.ts
@@ -1,3 +1,5 @@
-export * from './guard/navigation-lock.guard';
 export * from './directive/navigation-lock.directive';
-export * from './interface/has-navigation-lock';
+export * from './guard/navigation-lock.guard';
+export * from './service/navigation-lock.service';
+export * from './service/navigation-lock-handle';
+export * from './providers/navigation-lock.provider';